drm: samsung: corrrect ppc when calculating DPU ACLK
One DSC encoder provides 1 ppc capability. If DSC is enabled, we need to
consider DSC encoder number as the ppc to calculate ACLK for panel.

Bug: 257423894
Test: play various resolution videos
(cherry picked from commit 240c7039e381b9fb6402be99445a62d85c956c5b)

+++ samsung/exynos_drm_decon.h
@@ -143,7 +143,7 @@ struct dpu_bts {
 	u32 max_disp_freq;
 	u32 prev_max_disp_freq;
 	u32 dvfs_max_disp_freq;
-	u64 ppc;
+	u32 ppc;
 	u32 ppc_rotator;
 	u32 ppc_scaler;
 	u32 delay_comp;
